issue #2040 - updates per PR feedback
1. coreCosObjectResourceCountThreshold is now consistently an int
2. systemExportImpl is now "fast" in our default config

+++ operation/fhir-operation-bulkdata/src/main/java/com/ibm/fhir/operation/bulkdata/config/ConfigurationAdapter.java
@@ -131,7 +131,7 @@ public interface ConfigurationAdapter {
      *
      * @return
      */
-    long getCoreCosObjectResourceCountThreshold();
+    int getCoreCosObjectResourceCountThreshold();
 
     /**
      * @implNote System value.
